Builders
F this s

+++ crates/grappe/src/text/meta.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone, Debug)]
 pub struct TextMeta {
-    bytes: u16,
-    lines: u16,
+    bytes: usize,
+    lines: usize,
 }
